Extract LazyBatchColumn
