+++ src/coprocessor/codec/batch/mod.rs
@@ -11,8 +11,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod lazy_column;
 mod rows;
 mod vector;
 
+pub use self::lazy_column::LazyBatchColumn;
 pub use self::rows::BatchRows;
 pub use self::vector::VectorValue;
